requirements.txt
langchain
langchain_community
langchain_groq
validators
pytube
unstructured
python-dotenv